# openapi/content-workflow.openapi.yaml
openapi: 3.1.0
info:
  title: Content Workflow API
  version: 0.1.0
  summary: MVP API for content ingestion, drafting, review, publishing, and analytics snapshots.
  description: |
    This specification covers the MVP workflow:
    ingestion -> normalization -> topic selection -> draft generation -> asset rendering -> review -> publishing -> metrics snapshots.

    Implementation constraints encoded in this spec:
    - `Draft.sourceRefs` is required and must contain at least one source reference before review submission.
    - `ChannelVariant` is a platform-specific content object and does not expose or reuse the draft body as the publish payload.
    - `POST /publish-tasks` and `POST /publish-tasks/{publishTaskId}/run` are idempotent and require `Idempotency-Key`.
    - `PerformanceMetric` records time-series snapshots with `capturedAt`; the API does not support in-place metric updates.
    - `RawItem.contentHash` and `NormalizedItem.dedupeKey` are both required to support raw-content and semantic deduplication separately.
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
  /sources:
    get:
      tags: [Sources]
      summary: List content sources
      parameters:
        - $ref: '#/components/parameters/StatusQuery'
        - in: query
          name: platform
          schema:
            type: string
      responses:
        '200':
          description: Source list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceListResponse'
    post:
      tags: [Sources]
      summary: Create a content source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSourceRequest'
      responses:
        '201':
          description: Source created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '409':
          $ref: '#/components/responses/Conflict'
  /sources/{sourceId}:
    get:
      tags: [Sources]
      summary: Get a single source
      parameters:
        - $ref: '#/components/parameters/SourceId'
      responses:
        '200':
          description: Source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Sources]
      summary: Update a source
      parameters:
        - $ref: '#/components/parameters/SourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSourceRequest'
      responses:
        '200':
          description: Updated source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          $ref: '#/components/responses/NotFound'
  /sources/{sourceId}/sync:
    post:
      tags: [Ingestion]
      summary: Trigger source sync
      parameters:
        - $ref: '#/components/parameters/SourceId'
      responses:
        '202':
          description: Sync queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAcceptedResponse'
...
  /topics:
    get:
      tags: [Topics]
      summary: List topics
      parameters:
        - in: query
          name: date
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/StatusQuery'
      responses:
        '200':
          description: Topic list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
  /topics/generate:
    post:
      tags: [Topics]
      summary: Generate daily topic candidates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateTopicsRequest'
      responses:
        '200':
          description: Generated topic candidates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
  /topics/{topicId}:
    get:
      tags: [Topics]
      summary: Get a topic
      parameters:
        - $ref: '#/components/parameters/TopicId'
      responses:
        '200':
          description: Topic
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Topics]
      summary: Update a topic
      parameters:
        - $ref: '#/components/parameters/TopicId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTopicRequest'
      responses:
        '200':
          description: Updated topic
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
  /topics/{topicId}/select:
    post:
      tags: [Topics]
      summary: Mark a topic as selected
      parameters:
        - $ref: '#/components/parameters/TopicId'
      responses:
        '200':
          description: Topic selected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
  /topics/{topicId}/reject:
    post:
      tags: [Topics]
      summary: Reject a topic
      parameters:
        - $ref: '#/components/parameters/TopicId'
      responses:
        '200':
          description: Topic rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SourceId:
      in: path
      name: sourceId
      required: true
      schema:
        type: string
        format: uuid
...
    TopicId:
      in: path
      name: topicId
      required: true
      schema:
        type: string
        format: uuid
...
    StatusQuery:
      in: query
      name: status
      schema:
        type: string
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    SourceType:
      type: string
      enum: [rss, social_api, manual]
    SourceStatus:
      type: string
      enum: [active, disabled]
    TopicStatus:
      type: string
      enum: [candidate, selected, rejected, archived]
...
    Source:
      type: object
      additionalProperties: false
      required: [id, name, type, platform, config, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/SourceType'
        platform:
          type: string
        config:
          type: object
          additionalProperties: true
        status:
          $ref: '#/components/schemas/SourceStatus'
        lastSyncedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSourceRequest:
      type: object
      additionalProperties: false
      required: [name, type, platform, config]
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/SourceType'
        platform:
          type: string
        config:
          type: object
          additionalProperties: true
        status:
          $ref: '#/components/schemas/SourceStatus'
    UpdateSourceRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        config:
          type: object
          additionalProperties: true
        status:
          $ref: '#/components/schemas/SourceStatus'
...
    Topic:
      type: object
      additionalProperties: false
      required: [id, topicDate, title, score, status, clusterPayload, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        topicDate:
          type: string
          format: date
        title:
          type: string
        angle:
          type: string
        summary:
          type: string
        score:
          type: number
        status:
          $ref: '#/components/schemas/TopicStatus'
        clusterPayload:
          type: object
          additionalProperties: true
        selectedBy:
          type: string
          format: uuid
        selectedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GenerateTopicsRequest:
      type: object
      additionalProperties: false
      required: [date, targetSummaryCount, targetDeepCount]
      properties:
        date:
          type: string
          format: date
        targetSummaryCount:
          type: integer
          minimum: 1
        targetDeepCount:
          type: integer
          minimum: 1
    UpdateTopicRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        angle:
          type: string
        summary:
          type: string
        status:
          $ref: '#/components/schemas/TopicStatus'
...
    JobAcceptedResponse:
      type: object
      additionalProperties: false
      required: [jobId, status]
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued]
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    SourceListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Source'
...
    TopicListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Topic'
...